+++ CHANGELOG.md
@@ -16,5 +16,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Make no the default option when removing mods from an instance
 - Log error when no way to open web browser/file explorer
 - Use user downloads folder for Flatpak browser downloads
+- Don't prompt to install Fabric API when user already has Quilt Standard Libraries
 
 ### Misc